# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FriendsYINSH API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type GameRoomStatus =
  (typeof GameRoomStatus)[keyof typeof GameRoomStatus];

export const GameRoomStatus = {
  waiting: "waiting",
  playing: "playing",
  finished: "finished",
} as const;

export interface GameRoom {
  id: string;
  status: GameRoomStatus;
  createdAt: string;
  /** @nullable */
  currentPlayer: string | null;
  /** @nullable */
  winner: string | null;
  /** @nullable */
  gameState: string | null;
}