# lib/api-zod/src/generated/types/gameRoom.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FriendsYINSH API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameRoomStatus } from "./gameRoomStatus";

export interface GameRoom {
  id: string;
  status: GameRoomStatus;
  createdAt: string;
  /** @nullable */
  currentPlayer: string | null;
  /** @nullable */
  winner: string | null;
  /** @nullable */
  gameState: string | null;
}